refactor: 격리방법 수정 및 인수테스트 리팩토링
