+++ build.gradle
@@ -22,6 +22,7 @@ dependencies {
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'io.rest-assured:rest-assured:4.2.0'
+    testImplementation 'com.google.guava:guava:12.0'
 
     runtimeOnly 'com.h2database:h2'
 }
